Add import fix for packaging module.
This extension is distributed as a module so it essentially has
top-level code beyond a simple version attribute.  We need an
import of this module to work without additional packages so
that we can access the __version__ attribute in setup.py.
Otherwise, source distributions will not work.  This commit adds
an ImportError hack that makes this possible.